# .kilo/worktrees/cooing-chameleon/Documentation/Phase24/PHASE24_OPENAPI_SPECIFICATION.yaml
openapi: 3.1.0
info:
  title: DigiTronics V2 API
  description: |
    DigiTronics V2 Enterprise ERP API
    
    ## Authentication
    This API supports multiple authentication methods:
    - **JWT Bearer Token**: Standard JWT authentication
    - **OAuth2**: Google and GitHub OAuth2 providers
    - **API Key**: For service integrations (planned)
    
    ## Rate Limiting
    - General API: 1000 requests per 15 minutes
    - Login endpoints: 20 requests per 15 minutes
    - MFA endpoints: 5 requests per 15 minutes
    
    ## Error Format
    All errors follow a standard format:
    ```json
    {
      "success": false,
      "error": "Error message",
      "data": null
    }
    ```
  version: 2.0.0
  contact:
    name: DigiTronics Support
    email: support@digitronics.com
servers:
  - url: http://localhost:3001
    description: Development server
  - url: https://api.digitronics.com
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
    
    OAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: /auth/google
          tokenUrl: /auth/google/callback
          scopes:
            openid: OpenID Connect
            profile: User profile
            email: User email
    
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key authentication (planned)
  schemas:
...
    Purchase:
      type: object
      properties:
        id:
          type: string
          format: uuid
        orderNumber:
          type: string
        supplierName:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/PurchaseItem'
        total:
          type: number
          format: float
        status:
          type: string
          enum: [pending, received, cancelled]
        createdAt:
          type: string
          format: date-time
    PurchaseItem:
      type: object
      properties:
        productId:
          type: string
        productName:
          type: string
        quantity:
          type: integer
        unitCost:
          type: number
          format: float
        total:
          type: number
          format: float
    Inventory:
      type: object
      properties:
        id:
          type: string
          format: uuid
        productId:
          type: string
        productName:
          type: string
        quantity:
          type: integer
        minStock:
          type: integer
        maxStock:
          type: integer
        location:
          type: string
        updatedAt:
          type: string
          format: date-time
    Customer:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        address:
          type: string
        createdAt:
          type: string
          format: date-time
...
    Employee:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        position:
          type: string
        department:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        createdAt:
          type: string
          format: date-time
...
paths:
...
  /api/v1/inventory:
    get:
      tags: [Inventory]
      summary: List inventory
      description: Returns paginated list of inventory items
      security:
        - BearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Inventory retrieved
        '401':
          description: Authentication required
    
    post:
      tags: [Inventory]
      summary: Create inventory item
      description: Add a new item to inventory
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Inventory'
      responses:
        '201':
          description: Inventory item created
        '400':
          description: Validation error
        '401':
          description: Authentication required
  /api/v1/inventory/stats:
    get:
      tags: [Inventory]
      summary: Get inventory statistics
      description: Returns inventory count and value
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Statistics retrieved
        '401':
          description: Authentication required
  /api/v1/inventory/{id}:
    get:
      tags: [Inventory]
      summary: Get inventory item by ID
      description: Returns a specific inventory item
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Inventory item retrieved
        '404':
          description: Item not found
    
    put:
      tags: [Inventory]
      summary: Update inventory item
      description: Update an inventory item
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Inventory'
      responses:
        '200':
          description: Inventory item updated
        '404':
          description: Item not found
    
    delete:
      tags: [Inventory]
      summary: Delete inventory item
      description: Delete an inventory item
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Inventory item deleted
        '404':
          description: Item not found
  /api/v1/customers:
    get:
      tags: [Customers]
      summary: List customers
      description: Returns paginated list of customers
      security:
        - BearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Customers retrieved
        '401':
          description: Authentication required
    
    post:
      tags: [Customers]
      summary: Create customer
      description: Create a new customer
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Customer'
      responses:
        '201':
          description: Customer created
        '400':
          description: Validation error
        '401':
          description: Authentication required
...
  /api/v1/employees:
    get:
      tags: [Employees]
      summary: List employees
      description: Returns paginated list of employees
      security:
        - BearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Employees retrieved
        '401':
          description: Authentication required
    
    post:
      tags: [Employees]
      summary: Create employee
      description: Create a new employee
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Employee'
      responses:
        '201':
          description: Employee created
        '400':
          description: Validation error
        '401':
          description: Authentication required
  /api/v1/purchases:
    get:
      tags: [Purchases]
      summary: List purchases
      description: Returns paginated list of purchase orders
      security:
        - BearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: supplierName
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Purchases retrieved
        '401':
          description: Authentication required
    
    post:
      tags: [Purchases]
      summary: Create purchase
      description: Create a new purchase order
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Purchase'
      responses:
        '201':
          description: Purchase created
        '400':
          description: Validation error
        '401':
          description: Authentication required
...