# .worktree_recovery_deploy/backend/routes/purchase.routes.js
const router = require('express').Router();
const ctrl = require('../controllers/purchase.controller');
const asyncHandler = require('../utils/asyncHandler');
const { requirePermissionIfAuth } = require('../middleware/authorize');

/**
 * @openapi
 * /api/v1/purchases/stats:
 *   get:
 *     tags: [Purchases]
 *     summary: Get purchase statistics
 *     description: Returns purchase count and totals
 *     security:
 *       - BearerAuth: []
 *     responses:
 *       200:
 *         description: Statistics retrieved
 *       401:
 *         description: Authentication required
 */
router.get('/stats', requirePermissionIfAuth('purchases.view'), asyncHandler(ctrl.getStats));

/**
 * @openapi
 * /api/v1/purchases:
 *   get:
 *     tags: [Purchases]
 *     summary: List purchases
 *     description: Returns paginated list of purchase orders
 *     security:
 *       - BearerAuth: []
 *     parameters:
 *       - name: search
 *         in: query
 *         schema:
 *           type: string
 *       - name: supplierName
 *         in: query
 *         schema:
 *           type: string
 *       - name: status
 *         in: query
 *         schema:
 *           type: string
 *       - name: page
 *         in: query
 *         schema:
 *           type: integer
 *       - name: limit
 *         in: query
 *         schema:
 *           type: integer
 *     responses:
 *       200:
 *         description: Purchases retrieved
 *       401:
 *         description: Authentication required
 *   post:
 *     tags: [Purchases]
 *     summary: Create purchase
 *     description: Create a new purchase order
 *     security:
 *       - BearerAuth: []
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             $ref: '#/components/schemas/Purchase'
 *     responses:
 *       201:
 *         description: Purchase created
 *       400:
 *         description: Validation error
 *       401:
 *         description: Authentication required
 */
router.get('/', requirePermissionIfAuth('purchases.view'), asyncHandler(ctrl.list));
router.post('/', requirePermissionIfAuth('purchases.create'), asyncHandler(ctrl.create));

/**
 * @openapi
 * /api/v1/purchases/{id}:
 *   get:
 *     tags: [Purchases]
 *     summary: Get purchase by ID
 *     description: Returns a specific purchase order
 *     security:
 *       - BearerAuth: []
 *     parameters:
 *       - name: id
 *         in: path
 *         required: true
 *         schema:
 *           type: string
 *     responses:
 *       200:
 *         description: Purchase retrieved
 *       404:
 *         description: Purchase not found
 *   put:
 *     tags: [Purchases]
 *     summary: Update purchase
 *     description: Update a purchase order
 *     security:
 *       - BearerAuth: []
 *     parameters:
 *       - name: id
 *         in: path
 *         required: true
 *         schema:
 *           type: string
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             $ref: '#/components/schemas/Purchase'
 *     responses:
 *       200:
 *         description: Purchase updated
 *       404:
 *         description: Purchase not found
 *   delete:
 *     tags: [Purchases]
 *     summary: Delete purchase
 *     description: Delete a purchase order
 *     security:
 *       - BearerAuth: []
 *     parameters:
 *       - name: id
 *         in: path
 *         required: true
 *         schema:
 *           type: string
 *     responses:
 *       200:
 *         description: Purchase deleted
 *       404:
 *         description: Purchase not found
 */
router.get('/:id', requirePermissionIfAuth('purchases.view'), asyncHandler(ctrl.getById));
router.put('/:id', requirePermissionIfAuth('purchases.edit'), asyncHandler(ctrl.update));
router.delete('/:id', requirePermissionIfAuth('purchases.delete'), asyncHandler(ctrl.remove));

module.exports = router;